chore: api improvements
