+++ src/classes/index.ts
@@ -1,3 +1,1 @@
 export * from './copilot';
-export * from './completion/completion-formatter';
-export * from './completion/completion-validator';
