+++ src/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './copilot';
 export * from './editor';
-export * from './classnames/selection-actions';
