+++ src/core/selection-actions/modify/handler.ts
@@ -8,7 +8,7 @@ import {
   MODIFY_WIDGET_PROMPT_CONTAINER_CLASS,
   MODIFY_WIDGET_PROMPT_SUBMIT_BUTTON_CLASS,
   MODIFY_WIDGET_REJECT_BUTTON_CLASS,
-} from '../../../constants';
+} from '../../../constants/classnames';
 import {
   EditorSelection,
   ModifyOptions,
@@ -105,7 +105,7 @@ export const handleModifySelection = async (
           break;
       }`,
         );
-      }, 2000);
+      }, 1200);
     });
 
     // Remove loading state
@@ -147,11 +147,7 @@ export const handleModifySelection = async (
       submitButton.disabled = false;
     }
 
-    if (options.onError) {
-      options.onError(error as Error);
-    } else {
-      console.error(error);
-    }
+    throw error;
   }
 };
