+++ src/index.ts
@@ -1,14 +1,22 @@
 import {Copilot} from './classes';
 import {registerCompletion, registerCopilot} from './core/completion/register';
 import {registerSelectionActions} from './core/selection-actions/register';
+import {buildContext} from './helpers/context';
 
-export {registerCopilot, registerCompletion, registerSelectionActions, Copilot};
+export {
+  registerCopilot,
+  registerCompletion,
+  registerSelectionActions,
+  buildContext,
+  Copilot,
+};
 
 export type {
   Monaco,
   StandaloneCodeEditor,
   RegisterCompletionOptions,
   CompletionRegistration,
+  Context,
   CopilotOptions,
   CustomCopilotModel,
   CustomCopilotModelConfig,
