+++ src/prompts/index.ts
@@ -1,0 +1,2 @@
+export * from './completion-prompt';
+export * from './prompt-base';
