+++ src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './completion';
-export * from './modify';
+export * from './context';
 export * from './copilot';
+export * from './modify';
 export * from './monaco';
 export * from './provider';
