+++ src/types/modify.ts
@@ -19,7 +19,6 @@ export type ModifyOptions = {
   endpoint: string;
   placeholder?: string;
   requestHandler?: ModifyRequestHandler;
-  onError?: (error: Error) => void;
 };
 
 export type ModifyRequestHandler = (
